# docs/swagger/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Code understanding platform with hybrid search and LLM-powered enrichments",
        "title": "Kodit API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/queue": {
            "get": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "List tasks in the queue",
                "tags": [
                    "queue"
                ],
                "summary": "List tasks",
                "parameters": [
                    {
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Results per page (default: 20, max: 100)",
                        "name": "page_size",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Filter by task type",
                        "name": "task_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.TaskListResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/middleware.JSONAPIErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/queue/{task_id}": {
            "get": {
                "security": [
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Get a task by ID",
                "tags": [
                    "queue"
                ],
                "summary": "Get task",
                "parameters": [
                    {
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.TaskResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/middleware.JSONAPIErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/middleware.JSONAPIErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:8080/api/v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "APIKeyAuth": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        },
        "schemas": {
...
            "dto.TaskAttributes": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "payload": {},
                    "priority": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
            "dto.TaskData": {
                "type": "object",
                "properties": {
                    "attributes": {
                        "$ref": "#/components/schemas/dto.TaskAttributes"
                    },
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
            "dto.TaskListResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/dto.TaskData"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/jsonapi.Links"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/jsonapi.Meta"
                    }
                }
            },
            "dto.TaskResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/dto.TaskData"
                    }
                }
            },
...
            "jsonapi.Links": {
                "type": "object",
                "properties": {
                    "first": {
                        "type": "string"
                    },
                    "last": {
                        "type": "string"
                    },
                    "next": {
                        "type": "string"
                    },
                    "prev": {
                        "type": "string"
                    },
                    "self": {
                        "type": "string"
                    }
                }
            },
            "jsonapi.Meta": {
                "type": "object",
                "additionalProperties": {}
            },
            "middleware.JSONAPIError": {
                "type": "object",
                "properties": {
                    "detail": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    }
                }
            },
            "middleware.JSONAPIErrorResponse": {
                "type": "object",
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/middleware.JSONAPIError"
                        }
                    }
                }
            }
        }
    }
}